requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
pydantic==1.10.13
pydantic_core==2.10.1
PyJWT==2.8.0
sniffio==1.3.0
SQLAlchemy==1.4.50
sqlalchemy2-stubs==0.0.2a37
sqlmodel==0.0.11
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
